fix: restructure integration tests and add local test runners
