+++ .github/workflows/ci-ui-tests.yaml
@@ -87,7 +87,7 @@ jobs:
           sudo apt-get install python3-dev -y
 
       - name: run automatic_setup_microk8s.sh
-        run: integration_tests/automatic_setup_microk8s.sh
+        run: integration_tests/scripts/automatic_setup_microk8s.sh
 
       - name: install dependencies
         working-directory: ui_tests
