+++ integration_tests/scripts/check_for_pods.sh
@@ -1,0 +1,5 @@
+
+    while [ "$(sudo microk8s kubectl get pod -n sc4snmp | grep "worker-trap" | grep Running | wc -l)" != "1" ] ; do
+        echo "Waiting for POD initialization..."
+        sleep 1
+    done
