+++ integration_tests/scripts/install_sck.sh
