+++ integration_tests/scripts/local_cluster.bash
